fix: 全面安全与性能优化（CRITICAL/HIGH/MEDIUM）
C++ 后端:
- RedisService: incrWithExpire 改用 Lua 脚本保证原子性，修复 TOCTOU 竞态
- ProtocolDispatcher: 用 shared_ptr<atomic> + mutex 替代 map 引用捕获，修复并发安全
- PasswordUtils: PBKDF2 迭代次数 10k→200k，随机密码改用 RAND_bytes
- WebSocketManager: broadcast/sendToUser 先快照再发送，消除持锁调用 send 死锁风险
- AlertEngine: recordId 从 int 改为 int64_t，匹配 BIGINT 避免截断
- AuthCache: JWT 黑名单 key 改用 SHA256 哈希节省 Redis 内存
- AuthFilter: 日志不再记录 Token 内容
- DeviceCache: map→unordered_map + 新增辅助索引，查找 O(n)→O(1)
- SqlHelper: buildInClause 添加 static_assert 防止字符串类型误用
- WebSocket.Controller: shell:open 增加认证上下文校验

React 前端:
- WebSocket Token 从 URL query 改为 Sec-WebSocket-Protocol 传递，防止日志泄露
- refreshAccessToken 增加 userResponse.ok 检查，防止静默失败
- 5xx 错误提前 return，消除 notification + message 双重提示
- 路由器用 useRef 稳定 pageMenus 引用，避免 createHashRouter 重建
- 添加全局 ErrorBoundary 防止 render 异常白屏
- 新增 usePermissions hook，同组件多权限检查仅构建一次 Set
- tabsSlice.removeTab 修复关闭最后 tab 时的越界访问

构建优化:
- CMake: MSVC Release 启用 /GL + /LTCG
- CI: cmake --build 添加 --parallel，添加 bun 依赖缓存

+++ server/modules/alert/AlertEngine.hpp
@@ -134,7 +134,7 @@ public:
                         std::lock_guard lock(triggerStatesMutex_);
                         auto it = triggerStates_.find(rule.id);
                         if (it != triggerStates_.end()) {
-                            it->second.recordId = static_cast<int>(recordId);
+                            it->second.recordId = recordId;
                         }
                     }
                 } catch (const std::exception& e) {
@@ -175,7 +175,7 @@ private:
     };
 
     struct RuleTriggerState {
-        int recordId = 0;                       // 告警记录 ID
+        int64_t recordId = 0;                   // 告警记录 ID（匹配 BIGINT）
         Json::Value triggerData;                // 触发时的数据快照
         std::chrono::system_clock::time_point triggeredAt;
         bool isActive = true;                   // 是否处于活跃状态
@@ -345,7 +345,7 @@ private:
             std::lock_guard lock(triggerStatesMutex_);
             for (const auto& row : result) {
                 int ruleId = row["rule_id"].as<int>();
-                int recordId = row["id"].as<int>();
+                int64_t recordId = row["id"].as<int64_t>();
 
                 // 仅加载当前还存在的规则
                 if (triggerStates_.find(ruleId) != triggerStates_.end()) continue;
@@ -607,7 +607,7 @@ private:
     void markTriggerState(int ruleId, const Json::Value& data, int64_t recordId) {
         std::lock_guard lock(triggerStatesMutex_);
         triggerStates_[ruleId] = {
-            static_cast<int>(recordId),
+            recordId,
             data,
             std::chrono::system_clock::now(),
             true
@@ -617,7 +617,7 @@ private:
     Task<void> checkAndTriggerRecovery(const CachedRule& rule,
                                         const Json::Value& data) {
         // 先在锁内读取状态，co_await 前释放锁（避免持锁跨协程挂起导致死锁）
-        int recordId = 0;
+        int64_t recordId = 0;
         bool shouldRecover = false;
         std::string reason;
         {
@@ -717,7 +717,7 @@ private:
         return std::any_of(results.begin(), results.end(), [](bool r) { return r; });
     }
 
-    Task<void> triggerRecovery(const CachedRule& rule, int recordId,
+    Task<void> triggerRecovery(const CachedRule& rule, int64_t recordId,
                                 const Json::Value& data,
                                 const std::string& reason) {
         // 按 rule_id 恢复所有活跃记录（包括孤儿记录）
